big_data_study/requirements.txt
mrjob
pyspark
snakebite
hdfs
thrift
hbase-thrift
happybase
PyHive